refactor(cli): split audio module into file
